# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OrangeTrack - Sistema de Gestão de Laranjas
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerList"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /customers/{id}/orders:
    get:
      operationId: getCustomerOrders
      tags: [customers]
      summary: Get purchase history for a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  # PRODUCTS
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_production, ready, delivered, cancelled]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # INVENTORY
...
  /inventory/entries:
    post:
      operationId: createInventoryEntry
      tags: [inventory]
      summary: Add a lot entry (incoming stock)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryEntryBody"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryMovement"
  /inventory/exits:
    post:
      operationId: createInventoryExit
      tags: [inventory]
      summary: Register a stock exit (sale or discard)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryExitBody"
      responses:
        "201":
          description: Created exit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryMovement"
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        cnpj:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        phone:
          type: string
        email:
          type: string
        preferredProducts:
          type: array
          items:
            type: string
        totalOrders:
          type: integer
        totalSpent:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, name, cnpj, address, city, state, phone, email, createdAt]
    CustomerList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        cnpj:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        phone:
          type: string
        email:
          type: string
        preferredProducts:
          type: array
          items:
            type: string
      required: [name, cnpj, address, city, state, phone, email]

    # Product
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        totalPrice:
          type: number
        paymentType:
          type: string
          enum: [cash, credit, pix, invoice]
        status:
          type: string
          enum: [pending, in_production, ready, delivered, cancelled]
        deliveryType:
          type: string
          enum: [delivery, pickup]
        deliveryAddress:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - orderNumber
        - customerId
        - customerName
        - productId
        - productName
        - quantity
        - unit
        - totalPrice
        - paymentType
        - status
        - deliveryType
        - createdAt
        - updatedAt
    OrderList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    CreateOrderBody:
      type: object
      properties:
        customerId:
          type: integer
        productId:
          type: integer
        quantity:
          type: number
        paymentType:
          type: string
          enum: [cash, credit, pix, invoice]
        deliveryType:
          type: string
          enum: [delivery, pickup]
        deliveryAddress:
          type: string
        notes:
          type: string
      required: [customerId, productId, quantity, paymentType, deliveryType]
    UpdateOrderBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_production, ready, delivered, cancelled]
        notes:
          type: string

    # Inventory
...
    InventoryMovement:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [entry, exit_sale, exit_discard]
        productId:
          type: integer
        productName:
          type: string
        lotCode:
          type: string
        quantity:
          type: number
        unit:
          type: string
        reason:
          type: string
        orderId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, type, productId, productName, lotCode, quantity, unit, createdAt]
    CreateInventoryEntryBody:
      type: object
      properties:
        productId:
          type: integer
        lotCode:
          type: string
        quantity:
          type: number
        unit:
          type: string
        expiryDate:
          type: string
          format: date
      required: [productId, lotCode, quantity, unit]
    CreateInventoryExitBody:
      type: object
      properties:
        inventoryItemId:
          type: integer
        quantity:
          type: number
        type:
          type: string
          enum: [exit_sale, exit_discard]
        reason:
          type: string
        orderId:
          type: integer
      required: [inventoryItemId, quantity, type]

    # Production
...