# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OrangeTrack - Sistema de Gestão de Laranjas
 * OpenAPI spec version: 0.1.0
 */
import type { OrderDeliveryType } from "./orderDeliveryType";
import type { OrderPaymentType } from "./orderPaymentType";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  orderNumber: string;
  customerId: number;
  customerName: string;
  productId: number;
  productName: string;
  quantity: number;
  unit: string;
  totalPrice: number;
  paymentType: OrderPaymentType;
  status: OrderStatus;
  deliveryType: OrderDeliveryType;
  deliveryAddress?: string;
  notes?: string;
  createdAt: Date;
  updatedAt: Date;
}